requirements.txt
pandas
Pillow
python-dotenv